refactor: Calendar View Body 리팩토링, Sub Component 로 분리
